FEATURE: Encrypt the message queue
The `cluster` and `shadow` databases each with there own cursor
and cache. Shadow attributes are now in the `shadow` database in the
`attributes` table that is exactly the same schema as the one in the
`cluster` database.

TODO
	- host keys for mq
	- encrypt the SMQ (need authentication, could actually be in the clear but signed)
	- loopback traffic should be signed (don't do it at the shipper layer)

BREAKING CHANGE: schema addition

```
mysql
CREATE database shadow
USE database shadow
CREATE TABLE attributes (
	Scope		enum ('global', 'os', 'environment', 'appliance', 'host'),
	Attr		varchar(128) NOT NULL,
	Value		text,
	ScopeID		int(11)
);
```